change to rtol since things blow up fast.

+++ timemachine/cpu_functionals/wrappers.cpp
@@ -295,6 +295,6 @@ declare_electrostatics<float>(m, "float");
 declare_lennard_jones<double>(m, "double");
 declare_lennard_jones<float>(m, "float");
 
-declare_integrator<float>(m, "float");
+declare_integrator<double>(m, "double");
 
 }
